+++ arcore-location-extension/src/main/kotlin/org/neidhardt/arlocation/Extensions.kt
@@ -1,0 +1,5 @@
+package org.neidhardt.arlocation
+
+internal fun Float.toRadians(): Double {
+	return Math.toRadians(this.toDouble())
+}