+++ arcore-location-extension/src/test/kotlin/org/neidhardt/arlocation/LocationArNodeTest.kt
@@ -3,7 +3,6 @@ package org.neidhardt.arlocation
 import com.google.ar.sceneform.math.Vector3
 import org.junit.Assert.assertEquals
 import org.junit.Test
-import org.neidhardt.arlocation.misc.geodeticCurve
 
 class LocationArNodeTest {
 
@@ -20,19 +19,22 @@ class LocationArNodeTest {
 	@Test
 	fun getScaleGradual() {
 		// arrange
+		val src = GlobalPosition(52.001, 13.001)
+		val dst = GlobalPosition(52.002, 13.002)
 		val testDirection = Vector3(1f, 1f, 1f)
+		val distance = GlobalPositionUtils.geodeticCurve(
+				src,
+				dst
+		).ellipsoidalDistance
 		// action
 		val result = LocationArNode.getScaleGradual(
 				testDirection,
-				GlobalPosition(52.001, 13.001),
-				GlobalPosition(52.002, 13.002)
+				src,
+				dst
 		)
 		// verify
 		assertEquals(
-				LocationArNode.getScaleFixedSize(testDirection) * ArLocationUtils.scaleFactorForDistance(geodeticCurve(
-						52.001, 13.001,
-						52.002, 13.002
-				).ellipsoidalDistance),
+				LocationArNode.getScaleFixedSize(testDirection) * ArUtils.scaleFactorForDistance(distance),
 				result,
 				0.1
 		)